rag_chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.21.0",
    "pinecone>=7.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]
